+++ youshan/contribution.py
@@ -3,4 +3,4 @@ from models import theGroup
 
 def contrib(msg):
     group = theGroup(msg)
-    
+    # TODO
