+++ youshan/storage.py
@@ -7,7 +7,8 @@ from utils import formatToday
 
 def saveOnDisk(msg):
     '''
-    backward compatible
+    Deprecated.
+    Just for backward compatible
     TODO data migration
     '''
     the_group = theGroup(msg)
@@ -43,6 +44,50 @@ def saveOnDisk(msg):
         f.write(msg_name + ' = ' + str(content) + '\n')
     print(msg_name + ' saved on disk.')
     return True
+
+
+
+def getCorpus(arg='today', puid: str=None):
+    '''
+    Deprecated.
+    Only use to get old data on local disk storage.
+
+    :param `today`: `list` return a list of dict like
+    {'serial': '164',
+    'user': 'ch',
+    'date': '20180615',
+    'time': 1528996039.0,
+    'type': 'Text',
+    'text': '回不来了，已经晚了'}
+    :param `raw_today`: `str` return everything people said today
+    :param `raw_alltime`: `str` return everything people said since day0
+    :param `raw_puid_today`: `str` return everything a certain dude said today
+    '''
+    with open('db', 'r') as f:
+        all_dict = f.read().split('\n')
+    all_dict.remove('')
+    # today for instance: `20180615`
+    today = formatToday()
+    corpus_today = [eval(item.split(' = ')[1])
+                    for item in all_dict if today in item.split(' = ')[0]]
+    if arg == 'today':
+        return corpus_today
+    elif arg == 'raw_today':
+        return ''.join(
+            [item['text'] for item in corpus_today if item['text'] is not None])
+    elif arg == 'raw_alltime':
+        corpus_all = [eval(item.split(' = ')[1]) for item in all_dict]
+        return ''.join(
+            [item['text'] for item in corpus_all if item['text'] is not None])
+    elif arg == 'puid_today':
+        return [m for m in corpus_today if m['user'] == getUserByPuid(puid)]
+    elif arg == 'raw_puid_today':
+        name = getUserByPuid(puid)
+        user_corpus = [item for item in corpus_today if item['user'] == name]
+        user_texts = [item['text']
+                      for item in user_corpus if item['text'] is not None]
+        return ''.join(user_texts)
+
 
 
 def persistize(msg):
